dseed: update the output of `list_ch_peers` in the textbox
The `lbrytools.list_ch_peers` method was fixed to work
without errors with a channel that doesn't have claims.

In that case, the output of `act_peers.list_ch_peers`,
which is used by `list_ch_peers`, is a dictionary with an empty value
in the `'lines'` key.

If this value is empty, we replace it with text that indicates
that no claims were found, so that the summary printed to the textbox
is more complete.