pages_list: add new checkboxes to `ListChClaimsPage`
Since the `lbrytools.list_ch_claims` method was updated
with new parameters, `create`, `height`, `release`, `sizes`, `fees`,
the method to display channel claims in the interface `act_list.list_ch_claims`
is also updated.

----

The interface is updated with more checkboxes to display the new fields
in the claim information.

The new checkboxes are to display creation block height and creation time,
block height and timestamp, release time, duration and size, and fees.
The checkbox for "blocks" is removed, as that is included
in block height and timestamp.

The updated methods are `block_list.setup_check_ch_list`
to create the checkboxes.

This requires new variables, so `variables.VarsListChClaims` is updated
with five new variables, and one variable is removed.

Also update the information text in `pages_lists.ListChClaimsPage`.