variables: rename variables from `_tr_` to `_sr_`, part 2
Since the modules and classes were renamed to use the name 'search'
instead of 'trending', the variables used in the interface
replace the infix `_tr_` with `_sr_`.

These are the spinboxes for the page number and the number of threads,
the entry boxes for searching text and tags, and the radiobutton
to choose the type of claim.