pages_search: new class with common code for `TrendPage` and `SearchPage`
The new `BaseSearch` class is used as the common class for `TrendPage`
and `SearchPage`, which build the interfaces to display trending claims
and searched claims.

`BaseSearch.set_stream_vars_false` sets the 'all' checkbox to `True`
and all other type of stream variables to `False`.
This method is used in `TrendPage.deact_tr_checks` and `SearchPage.deact_sr_checks`.

`BaseSearch.switch_vars_all` sets the type of stream variables to `False`
if the 'all' checkbox is checked.
If the 'all' checkbox is not active, it will only set the checkboxes for videos
and documents to `True`.
This method is used in `TrendPage.switch_tr_all` and `SearchPage.switch_sr_all`.

`BasedSearch.switch_various` sets the 'all' checkbox to `False`
if any of the stream type checkboxes is set to `True`,
or sets the 'all' checkbox to `True` if all other stream type checkboxes
are set to `False`.
This method is used in `TrendPage.switch_tr_various` and `SearchPage.switch_sr_various`.

The output of the checkbox creation method `blocks.setup_check_trend_typ`
is stored in a tuple `self.tr_checks_typ` in `TrendPage`, and `self.sr_checks_typ`
in `SearchPage`.
Then these tuples are used to disable the widgets in `TrendPage.deact_tr_checks`
and `SearchPage.deact_sr_checks`, and to enable the widgets in `TrendPage.activate_tr_checks`
and `SearchPage.activate_sr_checks`.