act_comments: use `gmtime` for time formatting instead of `localtime`
In `show_comment` and `show_no_comment` instead of using `time.localtime`
to display the time of the claim we use `time.gmtime` so that the time
is displayed the same everywhere in the world.

We fix a bug when displaying the time.
Normally we try to display `'release_time'`, but since it was mispelled
as `'releaset_time'` it was displaying `'timestamp'` instead.
We correct the mispelling, and now if `'release_time'` is not available
we display `'creation_timestamp'`.

To format the time string we use the generic variable `lbrytools.funcs.TFMT`
instead of using a manual string such as `"%Y-%m-%d_%H:%M:%S%z %A"`.