lbrytools: update the submodule for `lbrytools/` to the latest version
Add new `check` method to quickly resolve and show information
on any claim.

It is based on the `search.search_item` method; after successfully
resolving the claim, it prints prints information about it
without downloading it.

This new `check` method could be used to show information on claims
in the interface.

----

Also add new method `get_download_dir` to check the passed directory
or get the default download directory saved in the configuration.

The directory where the files will be placed is only created
if the claim is of type `'stream'`.

----

A few smaller changes that don't impact `lbrydseed` such as renaming
of module files, and more compact code.