resolve: use the new `get_download_dir` method from `lbrytools`
The `lbrytools` library was updated with a new method `get_download_dir`
that checks if the passed directory actually exists, and if it doesn't
then it returns the default download directory specified by the `lbrynet`
configuration.

This means that for `lbrydseed` the old methods `get_download_dir`
and `check_download_dir`, defined in `resolve.py`, are no longer necessary.
The `check_download_dir` method is removed, and in its place
`get_download_dir` from `lbrytools` is used.

The `entry_d_dir` variable that specifies the initial value
of the download directory is also updated in `VarsDownload`
in `variables.py`.

The `server_exists` method in `resolve.py` is also replaced
with the corresponding method from `lbrytools`.