validate: re-organize the `validate_input` method in `validate.py`
The output of `validate_input` is a list of dictionaries.
Each dictionary has two keys, the first is renamed from `'claim'`
to `'claim_input'`, as it can be a name or claim ID.
The second one `'number'` remains the same.

Now the separator expected to parse the text can be specified
as one of the arguments to `validate_input`; it is set to a comma `,`
by default, as it was before.

----

Update the methods in `resolve.py` that use the information
from `validate_input`, that is, `i_resolve_chs` and `i_resolve_claims_supp`.

These methods are used in the `Download`, `Peers`, `Delete`,
and `Supports` pages.

Also update the list `validated_chs` (which has a single element)
which is used in `resolve_sg_ch`.
This is used in the `List` page (`list_ch_claims`) to print the claims
belonging to a single channel, and in the `Peers` page (`list_ch_peers`)
to list the peers in the claims of a single channel.