resolve: use `lbrytools.check` to check the input channel claims
Previously inside `resolve.i_resolve_chs` we would call
the `lbrynet resolve` method directly to resolve the claims.
Now we use the `lbrytools.check` method to resolve the claim
and also get the summary information for it.

The input channel is resolved twice, once by canonical URL,
and if not found, by claim ID.
This means that now the input channel could be given by claim ID
(40 character string) in addition to by channel `@name`, as before.

----

The output dictionary is updated; the name of the searched channel
is stored in `'claim_input'`, which may be a claim ID or name,
while `'claim'` has the fully resolved dictionary of the claim.
The `'summary'` key will hold textual information on the claim
which can be printed to the terminal or displayed
on another graphical interface.

The `'info'` key was removed because we don't really need it;
the same information, the `'canonical_url'`, can be taken
from the `'claim'` key.

----

Update the methods which use the output of `i_resolve_chs`
such as `act_download.i_download_chs`, `act_peers.i_list_chs_peers`,
`act_delete.i_delete_chs`, and also the main command `resolve_sg_ch`,
and thus `list_ch_claims` and `list_ch_peers`.