validate: the separator is changed to semicolon `;` by default
When processing the text with `validate_input` we set the separator
by default to `;`, instead of a comma `,` like before.

This is because claims can contain a comma as part
of their name like `some-claim,-with-a-comma`.

----

We also update the instances where the comma was used.

In `blocks_base.py` update `set_up_default_channels`
using a semicolon `;` as separator.

In `pages_support.py` update `setup_textbox_add_support`
using a semicolon `;` as separator.

In the `Download` (`page_down.py`), `Peers` (`pages_peers.py`) pages,
`Delete` (`pages_del.py`), and `Supports` (`pages_support.py`)
the text is changed to indicate that the semicolon is now the separator.