resolve: use `lbrytools.check` to check the input claims
Previously inside `resolve.i_resolve_claims` we would call
`lbrytools.search_item` to resolve the claims by canonical URL,
or by claim ID.
Now we use the `lbrytools.check` method to resolve the claim
and also get the summary information for it.

The output is updated; previously it was just a list of dictionaries,
where each dictionary was the resolved claim returned by `lbrytools.search_item`.
Now the dictionary has three keys, `'claim_input'` with the name of the claim
or claim ID used as input, `'claim'` with the fully resolved dictionary
as before, and `'summary'` with the textual information on the claim
which can be printed to the terminal or displayed
on another graphical interface.

----

Update the methods which use the output of `i_resolve_claims`
such as `act_download.i_download_claims`, the main command  `list_comments`,
`act_peers.i_list_m_peers`, and `act_delete.i_delete_claims`.

----

Previously in the internal command `act_download.i_download_claims`
if the claim did not exist, it would be skipped completely.

Now at least a message will be printed saying that the claim isn't valid.

----

Previously in the internal command `act_delete.i_delete_claims`
if the claim did not exist, it would be skipped completely.

Now at least a message will be printed saying that the claim isn't valid.

----

Previously in the internal command `act_peers.i_list_m_peers`
if the claim did not exist, it would be skipped completely.

Now at least a message will be printed saying that the claim isn't valid.

Also, for these invalid claims, a dictionary is added with only two keys,
`'resolved'` and `'original'`, so that the the internal method `lbrytools.list_m_peers`
can process them.