resolve: use `lbrytools.check` to check the input claims for supports
Previously inside `resolve.i_resolve_claims_supp` we would call
`lbrytools.search_item` to resolve the claims by canonical URL,
or by claim ID.
Now we use the `lbrytools.check` method to resolve the claim
but also get the summary information for it.

The output is updated; `'claim_input'` has the name of the claim
or claim ID used as input, `'claim'` has the fully resolved dictionary,
and `'summary'` has the textual information on the claim
which can be printed to the terminal or displayed
on another graphical interface.

The `'resolved'` key was removed because it is the same
as the new `'claim'` key. The `'number'` key remains the same.

----

Update the methods which use the `i_resolve_claims_supp` output
such as `resolve_g_claim`, in the main interface, only if the "invalid",
checkbutton is off.

In turn the output of `resolve_g_claim` is used by `update_supports`
which internally calls `act_supports.i_update_supports`.

----

The code of `act_supports.i_update_supports` is modified to consider
three cases, the `claim` exists, the `claim` doesn't exist and the "invalid"
checkbox is selected, and the `claim` doesn't exist and the "invalid"
checkbox isn't selected.