act_peers: rename the internal `list_subs_peers`, add an `i_` prefix
The new name is `i_list_subs_peers`.
This is to indicate that this is an internal method.

Also rename the two parameter names `ch_threads` and `c_threads`
to `ch_thrs` and `c_thrs`, respectively, just so the method call
of `i_list_subs_peers` isn't too long.

This main metod is renamed from `list_subs_peers` to `list_ch_subs_peers`
to match the corresponding command in `lbrytools`.