act_supports: rename the internal `add_supports`, add an `i_` prefix
The new name is `i_update_supports`.

This is to indicate that this is an internal method,
and that the support can be added or removed in different ways,
as it calls one of `lbrytools.create_support`, `lbrytools.abandon_support`,
`lbrytools.abandon_support_inv`, or `lbrytools.target_support`.

Correspondingly the method in the main interface `dseed.py` is renamed
from `add_supports` to `update_supports`.