pages_search: new checkbox to display the LBC that supports the claim
Since the methods of `lbrytools` were updated with a new `supports`
parameter
- `lbrytools.list_trending_claims`
- `lbrytools.list_search_claims`

a new checkbox is added to the `Trending claims` and `Search` interfaces
to optionally display the total amount of LBC supports on the listed claims.

The interface is defined by the classes `pages_search.TrendPage`
and `pages_search.SearchPage`.
The new checkbox is added in the auxiliary method that is common
to both, `blocks_search.setup_check_trend_fields`.

A new variable `chck_sr_supp` is added to `variables.VarsSearching`,
and it is passed to `act_search.list_trending` and `act_search.list_search`,
when called from the methods `list_trending_claims` and `list_search`
defined in the main interface `dseed.py`.