pages_list: new checkbox to display the LBC that supports the claim
Since the method of `lbrytools` was updated with a new `supports`
parameter
- `lbrytools.list_ch_claims`

a new checkbox is added to the `Channel claims` interface to optionally display
the total amount of LBC supports on the listed claims.

The interface is defined by the class `pages_lists.ListChClaimsPage`.
The new checkbox is added in the auxiliary method `blocks_list.setup_check_ch_list`.

A new variable `chck_chl_supp` is added to `variables.VarsListChClaims`,
and it is passed to `act_list.list_ch_claims` when called
from the method `list_ch_claims` defined in the main interface `dseed.py`.