act_cfg: rename the internal `list_lbrynet_settings`, add an `i_` prefix
This is to distinguish the method from the one defined
in the main interface `dseed.py`.

This main metod is renamed from `get_lbry_settings` to `list_lbrynet_settings`
to match the corresponding command in `lbrytools`.

Also adjust the command called by the button in the interface `pages_base.SettingsPage`.