dseed: rename the method `resolve_clm_cmnt` to `resolve_claim_cmnt`
This method calls `resolve.resolve_claims` to parse the input text
to get a claim for which to display comments.

In `resolve.resolve_claims` the `repost` parameter is set fixed to `True`,
instead of depending on the variable `check_d_repost`, which should be used
for the `Download` methods only.

The `pages_comments.CommentsPage` class is updated to call
the new `resolve_claim_cmnt` method.

----

The internal `act_comments.list_comments` method has some changes
to its code.
The output dictionary key `'text'` is renamed to `'summary'`,
as it's a summary of the total number of comments and replies
that the claim has.