dseeds: rename `resolve_ch` to `resolve_chs`
This method is used to resolve the input of channels
placed inside the text box, so it's better to be plural.

It affects the "Resolve online" button in three pages `Download`,
`Peers` and `Delete`, and also the methods `download_ch`,
`list_chs_peers`, and `delete_ch` defined in the main interface `dseed.py`.