dseed: rename `resolve_ch_list` to `resolve_sg_ch`
This method is used to resolve a single channel given
by a channel entry field.

The older name `resolve_ch_list` is confusing, as it sounds like
it resolves an entire list.

The renaming affects the "Resolve online" button in two pages `Channel claims`,
and `Channel peers`, and also the methods `list_ch_claims` and `list_ch_peers`
defined in the main interface `dseed.py`.