resolve: fix a bug when the number of claims is 0
If the number next to the channel is zero, it will be ignored
as if the number was set to `False` or `None`.

Now we change the test to test for `None` explicitly,
so that even numerical values of zero `0` are accepted.

Inside `resolve_ch_list`, the `validated_chs` variable is updated
so that now the `number` key is `None` instead of `False`
so tha the `resolve.resolve_ch` method can accept it well.