act_download: channel download/clean/peers only if they actually exist
Previously in the internal command `act_download.i_download_chs`
the method `lbrytools.ch_download_latest` would accept `number=0`,
which would default to `number=2`.

Now if `number=0` it will simply skip the channel
and not download anything.

If the channel was not successfully resolved
it will be skipped as well.

----

The `act_delete.i_delete_chs` method is also modified.

Previously if the channel did not exist, it would be skipped
completely.

Now at least a message will be printed saying that the channel isn't valid.

----

The `act_peers.i_list_chs_peers` method is also modified.

Previously if the channel did not exist, it would be skipped
completely.

Now every channel will apper in the output, even those which don't exist,
which will be marked by a pair of surrounding `[brackets]`.

The information text in the "Multiple channel peers" page (`ListChsPeersPage`)
is updated to explain this new behavior.