dseed: rename the variable `channels` to `resolved_chs` as they are dictionaries
The list of fully resolved channels is taken from `list_pub_chs`
which calls internally `act_list.i_list_pub_chs`
which calls `lbrytools.list_channels`.

This list is looped and used to extract the canonical name
of each individual channel, that is, `@channel#1af`.

The string values are used in the interface:
- `fill_ch_list` to update the combo box `combo_pub_ch`,
  to list the channels in the current wallet in the "Published claims" page
  (`ListPubClaimsPage`).

- `fill_ch_comment` to update the combo box `combo_author_cmnt`,
  to list the channels that can be used to write comments
  by calling the `reply_actions` method in the "Comments" page
  (`CommentsPage`).