dseed: method to disable editing of output textbox
The majority of the methods output information to a textbox.
This information doesn't need to be modified after being printed
so the textbox is set to be read-only.

We introduce the method `write_text` to make the given textbox editable
(`"normal"`), then write the output information to it,
and finally disable it to make it read-only (`"disabled"`).

This method is used for every output textbox for each page:
- "Settings" (`SettingsPage`)
- "Status" (`StatusPage`)
- "Downloaded claims" (`ListDownPage`)
- "Invalid claims" (`ListDownInvalidPage`)
- "Channel claims" (`ListChClaimsPage`)
- "Subscribed channels" (`SubscribedChsPage`)
- "Created channels" (`ListPubChsPage`)
- "Published claims" (`ListPubClaimsPage`)
- "Controlling claims" (`ControllingClaimsPage`)
- "Comments" (`CommentsReplyPage` and `CommentsPage`)
- "Claim peers" (`ListClsPeersPage`)
- "Channel peers" (`ListChPeersPage`)
- "Multiple channel peers" (`ListChsPeersPage`)
- "Subscription peers" (`ListSubsPeersPage`)
- "Seeding ratio" (`SeedPage`)
- "List supports" (`SupportListPage`)
- "Trending claims" (`TrendPage`)
- "Search" (`SearchPage`)

All these textboxes are set to `"disabled"` state by default.