pages_lists: rename the texboxes `textbox_list` to `textbox_list_d`
The infix `_d_` or `_down_` is meant to indicate "download"
as these textboxes are meant to display lists of downloaded claims,
both valid or invalid.

Other variables are renamed so that the infix `_d_` or `_down_` is after
the word `list`; for example, `page_down_list` is now `page_list_down`.