dseed: change the output of `resolve_sg_ch`
The output of `resolve_sg_ch` previously was a string with the resolved
channel name, such as `@channel#3`.

Now it is the full dictionary returned by `resolve.i_resolve_chs`.

Since this `resolve_sg_ch` method is used in a single channel,
the input list has only one element, and the output list
has only a single element as well.

If the resolved channel exists we write the resolved information
into the appropriate textbox for the "Channel claims" page (`ListChClaimsPage`)
or "Channel peers" page (`ListChPeersPage`).