pages_lists: add a status bar with information on the resolved claim
A new label is added to the bottom of the "Channel list" page
(`ListChClaimsPage`). Its content is controlled by a new variable,
`lab_ch_claims_status`, defined in `variables.VarsListChClaims`.

If the resolved channel is valid, the resolved name will be
placed in the status bar. If the channel isn't found,
it will also be indicated in the bar.

A similar bar is added to the "Channel peers" page (`ListChPeersPage`).
Its content is controlled by the new variable `lab_ch_peers_status`
defined in `variables.VarsPeers`.