pages_down: new textbox to place the resolved information of the claims
Various claims and channels can be processed at the same time
so we add a textbox to show basic information after resolving them,
including the claims that don't exist.

The new textbox is added to:
- "Download channels" (`DownloadChsPage`)
- "Download claims" (`DownloadClaimsPage`)

The `resolve_chs` and `resolve_claims` methods are updated
to place the resolved summary of each claim inside
the corresponding textbox.