pages_comments: new textbox to place the resolved information of the claims
Various claims can be processed at the same time
so we add a textbox to show basic information after resolving them,
including the claims that don't exist.

The new textbox is added to:
- "Comments" (`CommentsPage`)

The `resolve_claims` method is updated to place the resolved summary
of each claim inside the new textbox.

----

Previously the "Comments" page would use its own method `resolve_claim_cmnt`
to process the list of claims, but now the existing method `resolve_claims`
is used, so `resolve_claim_cmnt` is removed.