dseed: new `resolve_claims_d` to resolve the claims for download
The `resolve_claims` method is a general method that extracts
the claims in the textboxes in three pages
- "Download claims" (`DownloadClaimsPage`)
- "Delete claims" (`DeleteClaimsPage`)
- "Claim peers" (`ListClsPeersPage`)

In each case, the text is passed to the internal method
`resolve.i_resolve_claims`.

The `i_resolve_claims` method accepts the parameter `repost`
in order to resolve claims that are reposts.
Currently this parameter is only controlled by a checkbox
in the "Download claims" page.
The other pages don't control this checkbox, so if it is left
unchecked in "Downlaod claims", the behavior may not be as expected
by the user in the other pages.

Therefore, a new method is defined `resolve_claims_d`
which will be used only for the "Download claims" page (`DownloadClaimsPage`).
In this method the `repost` parameter is controlled by the checkbox.
Other pages will use the generic method `resolve_claims`
in which `repost=True` by default.