pages_support: new textbox to place the resolved information of the claims
Various claims and channels can be processed at the same time
so we add a textbox to show basic information after resolving them,
including the claims that don't exist.

The new textbox is added to:
- "Add or remove support" (`SupportUpdatePage`)

The `resolve_g_claims` method is updated to place the resolved summary
of each claim inside the new textbox.